Fixed error in console of update header by adding conditional logic
