+++ src/pages/index.tsx
@@ -1,73 +1,45 @@
 import type { NextPage } from "next"
-import Head from "next/head"
-import Image from "next/image"
-import styles from "../styles/Home.module.css"
+import { useState, useEffect } from "react"
+import jwtDecode from "jwt-decode"
+import AuthPage from "../components/authPage"
+import HouseDisplay from "../components/displayHouses"
 
 const Home: NextPage = () => {
-  return (
-    <div className={styles.container}>
-      <div className="font-sans text-3xl underline">Boom Shakalaka</div>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
+  const [isAuthorized, setIsAuthorized] = useState<boolean>(false) 
+  
+  useEffect(() => {
+      const token = localStorage.getItem('user-token')
+      const user = token ? jwtDecode(token) : null
+      if (user) {
+        setIsAuthorized(true)
+      }
 
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
+      window.addEventListener('storage', () => {
+        const token = localStorage.getItem('user-token')
+        const user = token ? jwtDecode(token) : null
+        if (user) {
+          setIsAuthorized(true)
+        }
+      })
 
-        <p className={styles.description}>
-          Get started by editing{" "}
-          <code className={styles.code}>pages/index.tsx</code>
-        </p>
+      return () => window.removeEventListener('storage', () => {
+        const token = localStorage.getItem('user-token')
+        const user = token ? jwtDecode(token) : null
+        if (user) {
+          setIsAuthorized(true)
+        }
+      })
+  }, [])
 
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
-
-          <a href="https://nextjs.org/learn" className={styles.card}>
-            <h2>Learn &rarr;</h2>
-            <p>Learn about Next.js in an interactive course with quizzes!</p>
-          </a>
-
-          <a
-            href="https://github.com/vercel/next.js/tree/canary/examples"
-            className={styles.card}
-          >
-            <h2>Examples &rarr;</h2>
-            <p>Discover and deploy boilerplate example Next.js projects.</p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-          >
-            <h2>Deploy &rarr;</h2>
-            <p>
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{" "}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
-    </div>
-  )
+  if (isAuthorized) {
+    return (
+      <HouseDisplay />
+    )
+  } else {
+    return (
+      <AuthPage />
+    )
+  }
 }
 
 export default Home
